[analytics-engine] Address review: dedicated IT, shared helper, cleanup
- Move shard plan tests from ExplainApiIT to dedicated PartialAggregateModeIT
- Extract runShardPlanCheck helper with mode, sliceCount, oversampling params
- Rename setConcurrentSearchMode (sets both mode and max_slice_count)
- Remove duplicate log_debug in Rust prepared-plan path
- Add TODO for consolidating partial-aggregate tree-walk methods
