+++ sandbox/qa/analytics-engine-rest/src/test/java/org/opensearch/analytics/qa/ExplainApiIT.java
@@ -304,142 +304,6 @@ public class ExplainApiIT extends AnalyticsRestTestCase {
             physicalPlan.contains("QueryShardExec"));
     }
 
-    /** slice=1, topk=off, vanilla scan */
-    @SuppressWarnings("unchecked")
-    public void testShardPlan_slice1_topkOff_vanilla() throws Exception {
-        ensureClickBenchProvisioned();
-        setSliceCount("none");
-        setOversampling(0.0);
-        assertShardPlanPartialOnly("source=" + CLICKBENCH.indexName + " | stats count() as c by RegionID | sort - c | head 10");
-    }
-
-    /** slice=1, topk=on, vanilla scan */
-    @SuppressWarnings("unchecked")
-    public void testShardPlan_slice1_topkOn_vanilla() throws Exception {
-        ensureClickBenchProvisioned();
-        setSliceCount("none");
-        setOversampling(2.0);
-        assertShardPlanPartialOnly("source=" + CLICKBENCH.indexName + " | stats count() as c by RegionID | sort - c | head 10");
-    }
-
-    /** slice>1, topk=off, vanilla scan */
-    @SuppressWarnings("unchecked")
-    public void testShardPlan_sliceN_topkOff_vanilla() throws Exception {
-        ensureClickBenchProvisioned();
-        setSliceCount("all");
-        setOversampling(0.0);
-        assertShardPlanPartialOnly("source=" + CLICKBENCH.indexName + " | stats count() as c by RegionID | sort - c | head 10");
-    }
-
-    /** slice>1, topk=on, vanilla scan */
-    @SuppressWarnings("unchecked")
-    public void testShardPlan_sliceN_topkOn_vanilla() throws Exception {
-        ensureClickBenchProvisioned();
-        setSliceCount("all");
-        setOversampling(2.0);
-        assertShardPlanPartialOnly("source=" + CLICKBENCH.indexName + " | stats count() as c by RegionID | sort - c | head 10");
-    }
-
-    /** slice=1, topk=off, delegation scan (filter on keyword) */
-    @SuppressWarnings("unchecked")
-    public void testShardPlan_slice1_topkOff_delegation() throws Exception {
-        ensureClickBenchProvisioned();
-        setSliceCount("none");
-        setOversampling(0.0);
-        assertShardPlanPartialOnly("source=" + CLICKBENCH.indexName + " | where URL != '' | stats count() as c by URL | sort - c | head 10");
-    }
-
-    /** slice=1, topk=on, delegation scan (filter on keyword) */
-    @SuppressWarnings("unchecked")
-    public void testShardPlan_slice1_topkOn_delegation() throws Exception {
-        ensureClickBenchProvisioned();
-        setSliceCount("none");
-        setOversampling(2.0);
-        assertShardPlanPartialOnly("source=" + CLICKBENCH.indexName + " | where URL != '' | stats count() as c by URL | sort - c | head 10");
-    }
-
-    /** slice>1, topk=off, delegation scan (filter on keyword) */
-    @SuppressWarnings("unchecked")
-    public void testShardPlan_sliceN_topkOff_delegation() throws Exception {
-        ensureClickBenchProvisioned();
-        setSliceCount("all");
-        setOversampling(0.0);
-        assertShardPlanPartialOnly("source=" + CLICKBENCH.indexName + " | where URL != '' | stats count() as c by URL | sort - c | head 10");
-    }
-
-    /** slice>1, topk=on, delegation scan (filter on keyword) */
-    @SuppressWarnings("unchecked")
-    public void testShardPlan_sliceN_topkOn_delegation() throws Exception {
-        ensureClickBenchProvisioned();
-        setSliceCount("all");
-        setOversampling(2.0);
-        assertShardPlanPartialOnly("source=" + CLICKBENCH.indexName + " | where URL != '' | stats count() as c by URL | sort - c | head 10");
-    }
-
-    @SuppressWarnings("unchecked")
-    private void assertShardPlanPartialOnly(String ppl) throws IOException {
-        Map<String, Object> result = executeWithProfile(ppl);
-        Map<String, Object> profile = (Map<String, Object>) result.get("profile");
-        assertNotNull("profile present", profile);
-        List<Map<String, Object>> stages = (List<Map<String, Object>>) profile.get("stages");
-
-        // Shard stage: only mode=Partial allowed
-        Map<String, Object> shardStage = stages.stream()
-            .filter(s -> "SHARD_FRAGMENT".equals(s.get("execution_type")))
-            .findFirst()
-            .orElseThrow(() -> new AssertionError("no SHARD_FRAGMENT stage"));
-
-        List<Map<String, Object>> tasks = (List<Map<String, Object>>) shardStage.get("tasks");
-        assertNotNull("shard tasks present", tasks);
-
-        for (Map<String, Object> task : tasks) {
-            Object plan = task.get("physical_plan");
-            if (plan instanceof String s && s.isEmpty() == false) {
-                assertFalse(
-                    "shard DF plan must NOT contain mode=Final. Plan:\n" + s,
-                    s.contains("mode=Final,") || s.contains("mode=Final ")
-                );
-                assertFalse(
-                    "shard DF plan must NOT contain FinalPartitioned. Plan:\n" + s,
-                    s.contains("FinalPartitioned")
-                );
-                assertTrue(
-                    "shard DF plan must contain mode=Partial. Plan:\n" + s,
-                    s.contains("mode=Partial")
-                );
-            }
-        }
-
-        // Coordinator reduce stage: must have Final or FinalPartitioned (merges shard partials)
-        stages.stream()
-            .filter(s -> "COORDINATOR_REDUCE".equals(s.get("execution_type")))
-            .findFirst()
-            .ifPresent(reduceStage -> {
-                String reducePlan = (String) reduceStage.get("physical_plan");
-                if (reducePlan != null && reducePlan.isEmpty() == false) {
-                    assertTrue(
-                        "coordinator plan must contain Final or FinalPartitioned. Plan:\n" + reducePlan,
-                        reducePlan.contains("mode=Final") || reducePlan.contains("FinalPartitioned")
-                    );
-                    assertFalse(
-                        "coordinator plan must NOT contain mode=Partial (shards handle that). Plan:\n" + reducePlan,
-                        reducePlan.contains("mode=Partial")
-                    );
-                }
-            });
-    }
-
-    private void setSliceCount(String mode) throws IOException {
-        Request req = new Request("PUT", "/_cluster/settings");
-        req.setJsonEntity("{\"transient\":{\"search.concurrent_segment_search.mode\":\"" + mode + "\"}}");
-        client().performRequest(req);
-    }
-
-    private void setOversampling(double factor) throws IOException {
-        Request req = new Request("PUT", "/_cluster/settings");
-        req.setJsonEntity("{\"transient\":{\"analytics.shard_bucket_oversampling_factor\":" + factor + "}}");
-        client().performRequest(req);
-    }
 
     @SuppressWarnings("unchecked")
     public void testProfileCoordinatorReduceHasMetrics() throws IOException {
